Reorganized, switched to SASS
