+++ package.json
@@ -15,7 +15,10 @@
   "devDependencies": {
     "@types/react": "^18.0.26",
     "@types/react-dom": "^18.0.9",
+    "@vitejs/plugin-legacy": "^3.0.1",
     "@vitejs/plugin-react-swc": "^3.0.0",
+    "sass": "^1.57.0",
+    "terser": "^5.16.1",
     "typescript": "^4.9.3",
     "vite": "^4.0.0"
   }
