+++ src/components/App.tsx
@@ -1,6 +1,6 @@
 import {useState} from 'react'
-import reactLogo from './assets/react.svg'
-import './App.css'
+import reactLogo from '../assets/react.svg'
+import '../styles/App.sass'
 
 function App() {
     const [count, setCount] = useState(0)
