+++ vite.config.ts
@@ -1,7 +1,14 @@
 import {defineConfig} from 'vite'
+import legacy from '@vitejs/plugin-legacy'
 import react from '@vitejs/plugin-react-swc'
 
 // https://vitejs.dev/config/
 export default defineConfig({
-    plugins: [react()],
+    plugins: [
+        react(),
+        legacy({
+            targets: ['defaults', 'not IE 11']
+        }),
+    ],
+    base: '/honors-portfolio/',
 })
